+++ src/components/news/NewsComment.jsx
@@ -1,4 +1,5 @@
 import { useState, useRef } from "react";
+import { formatDate } from "@/lib/utils";
 import send from "@/assets/send.svg";
 import deleteIcon from "@/assets/delete.svg";
 
@@ -30,23 +31,6 @@ export default function NewsComment({ comments }) {
   const onClickDelete = (commentId) => {
     console.log("click delete");
     setcommentList(commentList.filter((it) => it.commentId !== commentId));
-  };
-
-  const formatDate = (date) => {
-    const today = new Date();
-    const createdDate = new Date(date);
-
-    const diffMs = today - createdDate;
-    const diffMin = Math.floor(diffMs / (1000 * 60));
-    const diffHours = Math.floor(diffMs / (1000 * 60 * 60));
-    const diffDays = Math.floor(diffMs / (1000 * 60 * 60 * 24));
-
-    if (diffMin < 1) return "방금 전";
-    if (diffMin < 60) return `${diffMin}분 전`;
-    if (diffHours < 24) return `${diffHours}시간 전`;
-    if (diffDays < 7) return `${diffDays}일 전`;
-
-    return `${createdDate.getMonth() + 1}월 ${createdDate.getDate()}일`;
   };
 
   return (
